+++ api/app/models/group.py
@@ -29,26 +29,80 @@ class GroupBase(IncludedSamples):  # pylint: disable=too-few-public-methods
     """Basic specie information."""
 
     group_id: str = Field(..., min_length=5)
-    display_name: str = Field(..., min_length=1)
+    display_name: str = Field(..., min_length=1, max_length=45)
+    description: str | None = None
 
 
 class OverviewTableColumn(BaseModel):  # pylint: disable=too-few-public-methods
     """Definition of how to display and function of overview table."""
 
-    hidden: bool = Field(False)
-    type: str
-    path: str
-    label: str
-    sortable: bool = Field(False)
-    filterable: bool = Field(False)
-    filter_type: str | None = Field(None)
-    filter_param: str | None = Field(None)
+    id: str = Field(..., description="Column id")
+    label: str = Field(..., description="Display name")
+    path: str = Field(..., description="JSONpath describing how to access the data")
+    # display params
+    hidden: bool = False
+    type: str = Field("string", description="Data type")
+    sortable: bool = False
+    filterable: bool = False
+    filter_type: str | None = None
+    filter_param: str | None = None
+
+
+VALID_COLUMNS = [
+    OverviewTableColumn(
+        id="sample_id",
+        label="Sample Id",
+        type="sampleid",
+        path="$.sample_id",
+        sortable=True,
+    ),
+    OverviewTableColumn(
+        id="tags",
+        label="Tags",
+        type="tags",
+        path="$.tags",
+    ),
+    OverviewTableColumn(
+        id="taxonomic_name",
+        label="Major species",
+        type="taxonomic_name",
+        path="$.species_prediction.scientific_name",
+        sortable=True,
+    ),
+    OverviewTableColumn(
+        id="qc",
+        label="QC",
+        path="$.qc_status.status",
+        sortable=True,
+    ),
+    OverviewTableColumn(
+        id="mlst",
+        label="MLST ST",
+        path="$.mlst.sequence_type",
+        sortable=True,
+        filterable=True,
+    ),
+    OverviewTableColumn(
+        id="profile",
+        label="Analysis profile",
+        path="$.profile",
+        sortable=True,
+        filterable=True,
+    ),
+    OverviewTableColumn(
+        id="cdate",
+        label="Date",
+        type="date",
+        path="$.created_at",
+        sortable=True,
+    ),
+]
 
 
 class GroupInCreate(GroupBase):  # pylint: disable=too-few-public-methods
     """Defines expected input format for groups."""
 
-    table_columns: List[OverviewTableColumn] = Field(...)
+    table_columns: List[OverviewTableColumn] = Field(description="Columns to display")
     validated_genes: Dict[ElementType, List[str]] | None = Field({})