+++ CHANGELOG.md
@@ -21,3 +21,4 @@
  - Complete rewrite of the project
  - Removed N novel cgmlst alleles from cgmlst qc card
  - Use data models from JASEN Pipleine Result Processing application
+ - Use pre-defined table columns in edit groups view